feat(ops-editor): move attendance controls into the attendance tab
Unit assignment and the Discord ping toggle lived in a deck card while the
per-role ping targets that toggle gates lived in the Attendance tab — a
staging split left over from the editor rebuild rather than a deliberate
one. The tab had to render prose telling the reader to go flip a switch in
the sidebar before the roles below it would do anything.

Both halves are panels of the Attendance tab now and the deck has no
attendance card. No state moved: every handler was already lifted in
page.tsx, and the card was already isHQ-gated exactly as the tab is, so no
one's access changes.
